+++ tests/media/constraints.rs
@@ -51,7 +51,7 @@ async fn audio_constraints_satisfies() {
     let mut constraints = MediaStreamSettings::new();
     let mut track_constraints = AudioTrackConstraints::new();
     track_constraints.device_id(audio_device.device_id());
-    constraints.audio(track_constraints.clone());
+    constraints.device_audio(track_constraints.clone());
 
     let media_manager = MediaManager::default();
     let mut tracks =
@@ -86,7 +86,7 @@ async fn both_constraints_satisfies() {
         let mut video_constraints = DeviceVideoTrackConstraints::new();
         video_constraints.device_id(video_device.device_id());
 
-        constraints.audio(audio_constraints);
+        constraints.device_audio(audio_constraints);
         constraints.device_video(video_constraints);
 
         constraints
@@ -96,7 +96,7 @@ async fn both_constraints_satisfies() {
     let tracks = media_manager.get_tracks(constraints.clone()).await.unwrap();
 
     let video_constraints = constraints.get_device_video().clone().unwrap();
-    let audio_constraints = constraints.get_audio().clone();
+    let audio_constraints = constraints.get_device_audio().clone();
 
     assert_eq!(tracks.len(), 2);
 
@@ -213,7 +213,7 @@ async fn multi_source_media_stream_constraints_build1() {
 #[wasm_bindgen_test]
 async fn multi_source_media_stream_constraints_build2() {
     let mut constraints = MediaStreamSettings::new();
-    constraints.audio(AudioTrackConstraints::new());
+    constraints.device_audio(AudioTrackConstraints::new());
 
     let constraints: Option<MultiSourceTracksConstraints> = constraints.into();
 
@@ -236,7 +236,7 @@ async fn multi_source_media_stream_constraints_build2() {
 #[wasm_bindgen_test]
 async fn multi_source_media_stream_constraints_build3() {
     let mut constraints = MediaStreamSettings::new();
-    constraints.audio(AudioTrackConstraints::new());
+    constraints.device_audio(AudioTrackConstraints::new());
     constraints.device_video(DeviceVideoTrackConstraints::new());
 
     let constraints: Option<MultiSourceTracksConstraints> = constraints.into();
@@ -260,7 +260,7 @@ async fn multi_source_media_stream_constraints_build3() {
 #[wasm_bindgen_test]
 async fn multi_source_media_stream_constraints_build4() {
     let mut constraints = MediaStreamSettings::new();
-    constraints.audio(AudioTrackConstraints::new());
+    constraints.device_audio(AudioTrackConstraints::new());
     constraints.display_video(DisplayVideoTrackConstraints::new());
 
     let constraints: Option<MultiSourceTracksConstraints> = constraints.into();
@@ -350,7 +350,7 @@ fn get_device_video_track_constraints() -> DeviceVideoTrackConstraints {
 #[wasm_bindgen_test]
 async fn multi_source_media_stream_constraints_build7() {
     let mut constraints = MediaStreamSettings::new();
-    constraints.audio(AudioTrackConstraints::new());
+    constraints.device_audio(AudioTrackConstraints::new());
     constraints.device_video(get_device_video_track_constraints());
 
     let constraints: Option<MultiSourceTracksConstraints> = constraints.into();
@@ -387,6 +387,71 @@ async fn multi_source_media_stream_constraints_build8() {
 
             assert!(has_video);
             assert!(!has_audio);
+        }
+        _ => unreachable!(),
+    };
+}
+
+// Make sure that MediaStreamConstraints{audio:display, video:display} =>
+// Display({audio:true, video:true})
+#[wasm_bindgen_test]
+async fn multi_source_media_stream_constraints_build9() {
+    let mut constraints = MediaStreamSettings::new();
+    constraints.display_audio(AudioTrackConstraints::new());
+    constraints.display_video(DisplayVideoTrackConstraints::new());
+
+    let constraints: Option<MultiSourceTracksConstraints> = constraints.into();
+
+    match constraints {
+        Some(MultiSourceTracksConstraints::Display(constraints)) => {
+            let has_video =
+                js_val_to_option(constraints.as_ref().get_video()).is_some();
+            let has_audio =
+                js_val_to_option(constraints.as_ref().get_audio()).is_some();
+
+            assert!(has_video);
+            assert!(has_audio);
+        }
+        _ => unreachable!(),
+    };
+}
+
+// Make sure that MediaStreamConstraints{audio:any, video:any} =>
+// DeviceAndDisplay({audio:true, video:true}, {audio:display, video:display})
+#[wasm_bindgen_test]
+async fn multi_source_media_stream_constraints_build10() {
+    let mut constraints = MediaStreamSettings::new();
+    constraints.display_audio(AudioTrackConstraints::new());
+    constraints.display_video(DisplayVideoTrackConstraints::new());
+    constraints.device_audio(AudioTrackConstraints::new());
+    constraints.device_video(get_device_video_track_constraints());
+
+    let constraints: Option<MultiSourceTracksConstraints> = constraints.into();
+
+    match constraints {
+        Some(MultiSourceTracksConstraints::DeviceAndDisplay(
+            device_constraints,
+            display_constraints,
+        )) => {
+            let has_device_video =
+                js_val_to_option(device_constraints.as_ref().get_video())
+                    .is_some();
+            let has_device_audio =
+                js_val_to_option(device_constraints.as_ref().get_audio())
+                    .is_some();
+
+            assert!(has_device_video);
+            assert!(has_device_audio);
+
+            let has_display_video =
+                js_val_to_option(display_constraints.as_ref().get_video())
+                    .is_some();
+            let has_display_audio =
+                js_val_to_option(display_constraints.as_ref().get_audio())
+                    .is_some();
+
+            assert!(has_display_video);
+            assert!(has_display_audio);
         }
         _ => unreachable!(),
     };
@@ -451,7 +516,7 @@ fn build_constraints(
     if let Some(audio) = audio_device {
         let mut track_constraints = AudioTrackConstraints::new();
         track_constraints.device_id(audio.device_id());
-        constraints.audio(track_constraints);
+        constraints.device_audio(track_constraints);
     }
     if let Some(video) = video_device {
         let mut track_constraints = DeviceVideoTrackConstraints::new();
@@ -479,9 +544,10 @@ async fn simultaneous_device_and_display() {
         let mut video_constraints = DeviceVideoTrackConstraints::new();
         video_constraints.device_id(video_device.device_id());
 
-        constraints.audio(audio_constraints);
+        constraints.device_audio(audio_constraints);
         constraints.device_video(video_constraints);
         constraints.display_video(DisplayVideoTrackConstraints::new());
+        constraints.display_audio(AudioTrackConstraints::new());
 
         constraints
     };
@@ -493,9 +559,10 @@ async fn simultaneous_device_and_display() {
         constraints.get_device_video().clone().unwrap();
     let display_video_constraints =
         constraints.get_display_video().clone().unwrap();
-    let audio_constraints = constraints.get_audio().clone();
+    let device_audio_constraints = constraints.get_device_audio().clone();
+    let display_audio_constraints = constraints.get_display_audio().clone();
 
-    assert_eq!(tracks.len(), 3);
+    assert_eq!(tracks.len(), 4);
 
     let (mut audio, mut video): (Vec<_>, Vec<_>) =
         tracks.into_iter().partition(|(track, _)| match track.kind() {
@@ -503,9 +570,22 @@ async fn simultaneous_device_and_display() {
             MediaKind::Video => false,
         });
 
-    let audio_track = audio.pop().unwrap().0;
-    assert_eq!(audio_track.kind(), MediaKind::Audio);
-    assert!(audio_constraints.satisfies(audio_track.as_ref()).await);
+    let display_audio_track = audio.pop().unwrap().0;
+    assert_eq!(display_audio_track.kind(), MediaKind::Audio);
+    assert!(
+        display_audio_constraints.satisfies(display_audio_track.as_ref()).await
+    );
+    assert_eq!(
+        display_audio_track.media_source_kind(),
+        MediaSourceKind::Display
+    );
+
+    let device_audio_track = audio.pop().unwrap().0;
+    assert_eq!(device_audio_track.kind(), MediaKind::Audio);
+    assert!(
+        device_audio_constraints.satisfies(device_audio_track.as_ref()).await
+    );
+    assert_eq!(device_audio_track.media_source_kind(), MediaSourceKind::Device);
 
     let display_video_track = video.pop().unwrap().0;
     assert_eq!(display_video_track.kind(), MediaKind::Video);